feat: reintroduce vurderingsgrunnlag endpoint in intern API with updated routing
