+++ app/src/main/kotlin/no/nav/aap/arenaoppslag/InternRoute.kt
@@ -1,8 +1,10 @@
 package no.nav.aap.arenaoppslag
 
+import io.ktor.http.*
 import io.ktor.server.request.*
 import io.ktor.server.response.*
 import io.ktor.server.routing.*
+import no.nav.aap.arenaoppslag.kontrakt.apiv1.VurderingsgrunnlagRequest
 import no.nav.aap.arenaoppslag.kontrakt.intern.InternVedtakRequest
 import no.nav.aap.arenaoppslag.kontrakt.intern.PerioderMed11_17Response
 import no.nav.aap.arenaoppslag.kontrakt.intern.PerioderResponse
@@ -10,6 +12,8 @@ import no.nav.aap.arenaoppslag.kontrakt.intern.SakStatus
 import no.nav.aap.arenaoppslag.kontrakt.intern.SakerRequest
 import no.nav.aap.arenaoppslag.kontrakt.modeller.Maksimum
 import no.nav.aap.arenaoppslag.service.InternService
+import no.nav.aap.arenaoppslag.service.PersonService
+import no.nav.aap.arenaoppslag.service.VurderingsgrunnlagService
 
 fun Route.perioder(internService: InternService) {
     route("/perioder") {
@@ -61,4 +65,17 @@ fun Route.saker(internService: InternService) {
     }
 }
 
+fun Route.vurderingsgrunnlag(
+    vurderingsgrunnlagService: VurderingsgrunnlagService,
+    personService: PersonService,
+) {
+    post("/person/vurderingsgrunnlag") {
+        val request: VurderingsgrunnlagRequest = call.receive()
+        val personId = personService.hentPersonId(request.personidentifikator)
+            ?: return@post call.respond(HttpStatusCode.NotFound, "Fant ikke personen i Arena")
 
+        val respons = vurderingsgrunnlagService.hentVurderingsgrunnlag(personId)
+            ?: return@post call.respond(HttpStatusCode.NotFound, "Fant ingen AAP-sak for personen i Arena")
+        call.respond(HttpStatusCode.OK, respons)
+    }
+}
